inference_api/requirements.txt
fastapi[standard]
uvicorn
huggingface_hub
python-dotenv